separated config classes into different modules
