+++ test_autoconf/test_config.py
@@ -3,6 +3,7 @@ from os import path
 import pytest
 
 import autoconf as ac
+import autoconf.named
 
 directory = path.dirname(path.realpath(__file__))
 
@@ -15,7 +16,7 @@ class MockClass:
 
 @pytest.fixture(name="label_config")
 def make_label_config():
-    return ac.conf.LabelConfig("{}/test_files/config/label.ini".format(directory))
+    return autoconf.named.LabelConfig("{}/test_files/config/label.ini".format(directory))
 
 
 class TestLabel:
